🚚 Moved `src` to `studienarbeit` (poetry package docs)
