+++ .gitignore
@@ -189,3 +189,6 @@ cython_debug/
 *.code-workspace
 
 # End of https://www.toptal.com/developers/gitignore/api/python,visualstudiocode
+
+### Custom ###
+data/replication
