+++ pyproject.toml
@@ -15,9 +15,6 @@ classifiers = [
 ]
 license = "MIT"
 readme = "README.md"
-packages = [
-  { include = "src" },
-]
 
 [tool.poetry.dependencies]
 python = "^3.8"
@@ -26,7 +23,7 @@ pandas = "^1.5.0"
 numpy = "^1.23.3"
 
 [tool.poetry.group.dev.dependencies]
-black = {extras = ["jupyter"], version = "^22.8.0"}
+black = { extras = ["jupyter"], version = "^22.8.0" }
 isort = "^5.10.1"
 mypy = "^0.981"
 pylint = "^2.15.3"
@@ -38,15 +35,15 @@ pytest-cov = "^4.0.0"
 jupyter = "^1.0.0"
 
 [tool.poetry.group.docs.dependencies]
-mkdocs = {extras = ["python"], version = "^1.4.0"}
+mkdocs = { extras = ["python"], version = "^1.4.0" }
 mkdocs-material = "^8.5.6"
-mkdocstrings = {extras = ["python"], version = "^0.19.0"}
+mkdocstrings = { extras = ["python"], version = "^0.19.0" }
 mkdocs-autorefs = "^0.4.1"
 mkdocs-git-revision-date-localized-plugin = "^1.1.0"
 mkdocs-jupyter = "^0.22.0"
 
 [tool.mypy]
-mypy_path = 'src'
+mypy_path = 'studienarbeit'
 check_untyped_defs = true
 disallow_any_generics = true
 ignore_missing_imports = true
